+++ modules/sfml/src/sf_module.cpp
@@ -23,7 +23,7 @@ Value sf_window_init(const std::vector<Value>& n)
     
     if (!has_window)
     {
-        window.create(sf::VideoMode(n[0].number().toLong(), n[1].number().toLong()), n[2].string());
+        window.create(sf::VideoMode(static_cast<long>(n[0].number()), static_cast<long>(n[1].number())), n[2].string());
         has_window = true;
     }
     else
@@ -132,13 +132,13 @@ Value sf_window_clear(const std::vector<Value>& n)
         throw Ark::TypeError("sf-window-clear: g must be a Number");
     if (!n[2].isNumber())
         throw Ark::TypeError("sf-window-clear: b must be a Number");
-    window.clear(sf::Color(n[0].number().toLong(), n[1].number().toLong(), n[2].number().toLong()));
+    window.clear(sf::Color(static_cast<long>(n[0].number()), static_cast<long>(n[1].number()), static_cast<long>(n[2].number())));
     return nil;
 }
 
 Value sf_draw(const std::vector<Value>& n)
 {
-    for (Value::Iterator it=n.begin()+1; it != n.end(); ++it)
+    for (Value::Iterator it=n.begin(); it != n.end(); ++it)
     {
         if (!it->isString())
             throw Ark::TypeError("sf-draw: invalid argument");
@@ -170,7 +170,7 @@ Value sf_window_set_fps(const std::vector<Value>& n)
 {
     if (!n[0].isNumber())
         throw Ark::TypeError("sf-window-setFPS: fps must be a Number");
-    window.setFramerateLimit(n[0].number().toLong());
+    window.setFramerateLimit(static_cast<long>(n[0].number()));
     return nil;
 }
 
@@ -229,11 +229,11 @@ Value sf_make_text(const std::vector<Value>& n)
     texts[name] = sf::Text();
     texts[name].setFont(fonts[n[0].string()]);
     texts[name].setString(n[1].string());
-    texts[name].setCharacterSize(n[2].number().toLong());
+    texts[name].setCharacterSize(static_cast<long>(n[2].number()));
     texts[name].setFillColor(sf::Color(
-        n[3].const_list()[0].number().toLong(),
-        n[3].const_list()[1].number().toLong(),
-        n[3].const_list()[2].number().toLong()
+        static_cast<long>(n[3].const_list()[0].number()),
+        static_cast<long>(n[3].const_list()[1].number()),
+        static_cast<long>(n[3].const_list()[2].number())
     ));
 
     return Value(name);
@@ -254,9 +254,9 @@ Value sf_setpos(const std::vector<Value>& n)
     std::string sub = n[0].string().substr(0, i);
 
     if (sub == "text")
-        texts[n[0].string()].setPosition(n[1].number().toLong(), n[2].number().toLong());
+        texts[n[0].string()].setPosition(static_cast<long>(n[1].number()), static_cast<long>(n[2].number()));
     else if (sub == "sprite")
-        sprites[n[0].string()].setPosition(n[1].number().toLong(), n[2].number().toLong());
+        sprites[n[0].string()].setPosition(static_cast<long>(n[1].number()), static_cast<long>(n[2].number()));
     else if (sub == "event")
         throw Ark::TypeError("sf-setPos: Can not set position of event");
     else if (sub == "font")
@@ -316,7 +316,7 @@ Value sf_event(const std::vector<Value>& n)
     if (n.size() == 0)
         throw std::runtime_error("sf-event need at least 1 argument");
     
-    for (Value::Iterator it=n.begin()+1; it != n.end(); ++it)
+    for (Value::Iterator it=n.begin(); it != n.end(); ++it)
     {
         if (!it->isString())
             throw Ark::TypeError("sf-event: invalid argument");